docs: support dark bg
